+++ HISTORY.md
@@ -68,6 +68,9 @@
     * Chinese, thanks, Simon!
     * Brazilian Portuguese, thank you, Cassiano D'Andrea!
 
+### 2021-12-05
+* blocks, threads, objects: introspection & syntax analysis, experimental
+
 ### 2021-12-03
 * blocks, threads: block-assembly support for multi-args, experimental
 * threads: changed the visualization for an empty context to reporter-ring